funciones del reproductor
-recuperar ultima canción/episodio sonando
-guardar ultima canción/episodio sonando
-mejor organización script db
