+++ turso-scripts/00_all.sql
@@ -2,7 +2,6 @@ DROP TABLE IF EXISTS Listas_de_carpeta;
 DROP TABLE IF EXISTS Carpetas_del_usuario;
 DROP TABLE IF EXISTS Carpeta;
 DROP TABLE IF EXISTS Listas_del_usuario;
-DROP TABLE IF EXISTS Usuario_reproduce;
 DROP TABLE IF EXISTS Mensaje;
 DROP TABLE IF EXISTS Token;
 DROP TABLE IF EXISTS Sigue_a_usuario;
@@ -203,7 +202,9 @@ CREATE TABLE Usuario (
     contrasena      VARCHAR(255) NOT NULL,
     correo          VARCHAR(255) NOT NULL,
     link_compartir  VARCHAR(255),
-    es_admin        BOOLEAN NOT NULL DEFAULT FALSE
+    es_admin        BOOLEAN NOT NULL DEFAULT FALSE,
+    ult_cm          INTEGER REFERENCES Contenido_multimedia(id_cm) ON DELETE CASCADE,
+    tiempo_ult_cm   TIME
 );
 
 CREATE TABLE Sigue_a_creador (
@@ -237,15 +238,6 @@ CREATE TABLE Mensaje (
     fecha                 DATETIME,
     FOREIGN KEY (nombre_usuario_envia) REFERENCES Usuario(nombre_usuario) ON DELETE CASCADE,
     FOREIGN KEY (nombre_usuario_recibe) REFERENCES Usuario(nombre_usuario) ON DELETE CASCADE
-);
-
-CREATE TABLE Usuario_reproduce (
-    nombre_usuario  VARCHAR(255),
-    id_cm           INTEGER,
-    tiempo          TIME,
-    PRIMARY KEY (nombre_usuario, id_cm),
-    FOREIGN KEY (nombre_usuario) REFERENCES Usuario(nombre_usuario) ON DELETE CASCADE,
-    FOREIGN KEY (id_cm) REFERENCES Contenido_multimedia(id_cm) ON DELETE CASCADE
 );
 
 CREATE TABLE Listas_del_usuario (
@@ -592,7 +584,6 @@ INSERT INTO Idiomas_multimedia (id_cm, idioma) VALUES
 (35, 'Español'),
 (36, 'Español');
 
-
 INSERT INTO Sigue_a_creador (nombre_usuario, nombre_creador) VALUES
 ('jorge', 'Cruz Cafuné'),
 ('jorge', 'Bad Bunny');
@@ -658,10 +649,12 @@ INSERT INTO Lista_reproduccion (nombre, es_publica, color, link_compartir) VALUE
 ('ENTREVISTAS', TRUE, '#008000', 'https://example.com/share/entrevistas');
 
 INSERT INTO Playlist (id_playlist) VALUES
-(1), (2), (3), (4), (5), (6), (7), (8), (9), (10), (11), (12), (13),
-(17), (18), (19), (20), (21), (22), (23), (24);
+(1), (2), (3), (4), (5), (6), (7), (8), (9), (10), (11), (12), (13), (14), (15), (16),
+(17), (18), (19), (20), (21), (22), (23), (24), (25), 
+(29), (30), (31), (32), (33), (34), (35), (36);
 
 INSERT INTO Lista_Episodios (id_lista_ep) VALUES
+<<<<<<< Updated upstream:create-tables-bd.sql
 (14), (15), (16), (25), (26), (27), (28), (29), (30), (31), (32), (33), (34);
 
 INSERT INTO Listas_de_carpeta (id_carpeta, id_lista) VALUES
@@ -700,3 +693,99 @@ INSERT INTO Listas_de_carpeta (id_carpeta, id_lista) VALUES
 (6, 33),
 (6, 34);
 
+=======
+(26), (27), (28), 
+(37), (38), (39), (40), (41), (42), (43), (44), (45), (46);
+
+INSERT INTO Canciones_en_playlist (id_playlist, id_cancion) VALUES
+-- TOP ESPAÑOL
+(1, 2), (1, 4), (1, 11), (1, 12), (1, 13), (1, 14), (1, 15), (1, 16), (1, 17), (1, 19), (1,20), (1, 22),
+-- TOP FRANCÉS
+(2, 5),
+-- TOP INGLÉS
+(3, 18),(3,21),
+-- RAP
+(4, 1),(4,5),(4,32),
+-- REGUETÓN
+(5, 2),(5,10),(5,11),(5,12),(5,13),(5,14),(5,15),(5,16),(5,19),(5,23),(5,24),(5,25),(5,31),(5,33),(5,34),
+-- TRAP
+(6, 3),(6,4),(6,17),(6,20),(6,22),(6,35),(6,36),
+-- POP
+(7, 18),(7,21),
+-- This is Cruz Cafuné
+(8, 1),
+-- This is Bad Bunny
+(9, 2),(9,10),(9,12),(9,13),(9,14),
+-- This is Lola Índigo
+(10, 11),(10,15),(10,23),(10,24),(10,25),
+-- This is Sosad.97
+(11, 3),
+-- This is Feid
+(12, 16),(12,19),
+-- This is Duki
+(13, 4),(13,17),(13,20),(13,22),
+-- This is Harry Styles
+(14, 18),(14,21),
+-- This is Bigflo & Oli
+(15, 5),
+-- This is Paulo Londra
+(16, 22),
+-- This is María Becerra
+(17, 24),
+-- This is Emilia
+(18, 40), (18, 41), (18, 42), (18, 23),
+-- This is Don Patricio
+(19, 23),
+-- This is Quevedo
+(20, 25),
+-- This is Dei V
+(21, 2),
+-- This is Omar Courtz
+(22, 2),
+-- This is Efecto Pasillo
+(23, 37), (23, 38), (23, 39),
+-- This is Rels B
+(24, 31),(24,32),(24,33),(24,34),(24,35),(24,36),
+-- This is C. Tangana
+(25, 26),(25,27),(25,28),(25,29),(25,30),
+-- VERANO
+(29, 10),(29,12),(29,20),(29,21),(29,25),
+-- FIESTA
+(30, 4),(30,12),(30,20),(30,21),(30,22),(30,23),(30,24),
+-- RELAX
+(31, 18),(31,19),(31,31), (31,35),
+-- ENTRENAMIENTO
+(32, 3),(32,4),(32,13),(32,16),(32,17),(32,20),(32,22),(32,24),
+-- TRABAJO
+(33, 1),(33,5),(33,18),(33,21),(33,31),
+-- ESTUDIO
+(33, 15),(34,26),(34,27),(34,28),(34,29),(34,30),(33,34),
+-- SAN VALENTÍN
+(35, 23),(35,24),(35,25),(35,35),
+-- NAVIDAD
+(36, 18);
+
+
+INSERT INTO Episodios_de_lista (id_lista_ep, id_ep, id_podcast) VALUES
+(26, 6, 1), (26, 7, 1),
+(27, 6, 1), (27, 7, 1),
+(28, 8, 2), (28, 9, 2),
+(37, 6, 1), (37, 7, 1), (37, 8, 2), (37, 9, 2),
+(38, 6, 1), (38, 7, 1), (38, 8, 2), (38, 9, 2),
+(39, 6, 1), (39, 7, 1), (39, 8, 2), (39, 9, 2),
+(40, 6, 1), (40, 7, 1), (40, 8, 2), (40, 9, 2),
+(41, 6, 1), (41, 7, 1), (41, 8, 2), (41, 9, 2),
+(42, 6, 1), (42, 7, 1), (42, 8, 2), (42, 9, 2),
+(43, 6, 1), (43, 7, 1), (43, 8, 2), (43, 9, 2),
+(44, 6, 1), (44, 7, 1), (44, 8, 2), (44, 9, 2),
+(45, 6, 1), (45, 7, 1), (45, 8, 2), (45, 9, 2),
+(46, 6, 1), (46, 7, 1), (46, 8, 2), (46, 9, 2);
+
+INSERT INTO Listas_de_carpeta (id_carpeta, id_lista) VALUES
+(1, 1), (1, 2), (1, 3),
+(2, 4), (2, 5), (2, 6), (2, 7),
+(3, 8), (3, 9), (3, 10), (3, 11),(3, 12),(3, 13),(3, 14),(3, 15),(3, 16),(3, 17),(3, 18),(3, 19),(3, 20),(3, 21),(3, 22),(3, 23),(3, 24),(3, 25),
+(4, 26), (4, 27), (4, 28),
+(5, 29), (5, 30), (5, 31), (5, 32), (5, 33), (5, 34), (5, 35), (5, 36),
+(6, 37), (6, 38), (6, 39), (6, 40), (6, 41), (6, 42), (6, 43), (6, 44),(6, 45), (6, 46);
+>>>>>>> Stashed changes:turso-scripts/00_all.sql
