+++ turso-scripts/01_delete-tables.sql
@@ -1,0 +1,31 @@
+DROP TABLE IF EXISTS Listas_de_carpeta;
+DROP TABLE IF EXISTS Carpetas_del_usuario;
+DROP TABLE IF EXISTS Carpeta;
+DROP TABLE IF EXISTS Listas_del_usuario;
+DROP TABLE IF EXISTS Mensaje;
+DROP TABLE IF EXISTS Token;
+DROP TABLE IF EXISTS Sigue_a_usuario;
+DROP TABLE IF EXISTS Sigue_a_creador;
+DROP TABLE IF EXISTS Usuario;
+DROP TABLE IF EXISTS Canciones_en_playlist;
+DROP TABLE IF EXISTS Playlist;
+DROP TABLE IF EXISTS Episodios_de_lista;
+DROP TABLE IF EXISTS Lista_Episodios;
+DROP TABLE IF EXISTS Lista_reproduccion;
+DROP TABLE IF EXISTS Tiene_podcast;
+DROP TABLE IF EXISTS Podcaster;
+DROP TABLE IF EXISTS Tematica_podcast;
+DROP TABLE IF EXISTS Podcast;
+DROP TABLE IF EXISTS Episodio;
+DROP TABLE IF EXISTS Featuring;
+DROP TABLE IF EXISTS Artista_principal;
+DROP TABLE IF EXISTS Generos;
+DROP TABLE IF EXISTS Cancion;
+DROP TABLE IF EXISTS Idiomas_multimedia;
+DROP TABLE IF EXISTS Numero_cancion_en_album;
+DROP TABLE IF EXISTS Artista_posee_albumes;
+DROP TABLE IF EXISTS Album;
+DROP TABLE IF EXISTS Valoraciones;
+DROP TABLE IF EXISTS Contenido_multimedia;
+DROP TABLE IF EXISTS Artista;
+DROP TABLE IF EXISTS Creador;
